+++ packages/tailwindcss-language-server/tests/hover/hover.test.js
@@ -214,7 +214,7 @@ withFixture('v4/basic', (c) => {
     text: '<div class="bg-red-500">',
     position: { line: 0, character: 13 },
     expected:
-      '.bg-red-500 {\n  background-color: var(--color-red-500) /* oklch(0.637 0.237 25.331) = #fb2c36 */;\n}',
+      '.bg-red-500 {\n  background-color: var(--color-red-500) /* oklch(63.7% 0.237 25.331) = #fb2c36 */;\n}',
     expectedRange: {
       start: { line: 0, character: 12 },
       end: { line: 0, character: 22 },
@@ -231,16 +231,15 @@ withFixture('v4/basic', (c) => {
     },
   })
 
-  test.todo('arbitrary value with theme function')
-  // testHover('arbitrary value with theme function', {
-  //   text: '<div class="p-[theme(spacing.4)]">',
-  //   position: { line: 0, character: 13 },
-  //   expected: '.p-\\[theme\\(spacing\\.4\\)\\] {\n' + '  padding: 1rem /* 16px */;\n' + '}',
-  //   expectedRange: {
-  //     start: { line: 0, character: 12 },
-  //     end: { line: 0, character: 32 },
-  //   },
-  // })
+  testHover('arbitrary value with theme function', {
+    text: '<div class="p-[theme(spacing.4)]">',
+    position: { line: 0, character: 13 },
+    expected: '.p-\\[theme\\(spacing\\.4\\)\\] {\n' + '  padding: 1rem /* 16px */;\n' + '}',
+    expectedRange: {
+      start: { line: 0, character: 12 },
+      end: { line: 0, character: 32 },
+    },
+  })
 
   testHover('arbitrary property', {
     text: '<div class="[text-wrap:balance]">',